Cleanup

+++ fabric/src/main/java/com/lx862/jcm/mod/render/gui/screen/base/TitledScreen.java
@@ -55,7 +55,7 @@ public abstract class TitledScreen extends AnimatedScreen {
         graphicsHolder.translate(width / 2.0, TEXT_PADDING, 0);
         graphicsHolder.translate(0, -((titleHeight + TEXT_PADDING) * (1 - animationProgress)), 0);
         graphicsHolder.scale(TITLE_SCALE, TITLE_SCALE, TITLE_SCALE);
-        RenderHelper.scaleToFit(graphicsHolder, GraphicsHolder.getTextWidth(titleText), width / TITLE_SCALE, true);
+        RenderHelper.scaleToFit(graphicsHolder, GraphicsHolder.getTextWidth(titleText), width / (float)TITLE_SCALE, true);
         graphicsHolder.drawCenteredText(titleText, 0, 0, 0xFFFFFFFF);
         graphicsHolder.pop();
     }
